feet: refractored firebase module retrieval
