+++ site/src/lib/search/firebase/index.ts
@@ -38,3 +38,6 @@ export const COLL_MODULES = 'modules';
 export const COLL_STAFF = 'staff';
 export const COLL_VENUES = 'venues';
 export const SUB_COLL_SEMESTERS = 'semesters';
+
+export { semesterIdToSemYear, getSemesterTitle } from './helper';
+export { getModuleDoc } from './module';
