+++ site/src/routes/search/firebase/module/+server.ts
@@ -1,18 +1,11 @@
-import { COLL_MODULES, db } from '$lib/search/firebase';
-import { json } from '@sveltejs/kit';
-import { collection, getDocs, query, where } from 'firebase/firestore';
+import { getModuleDoc, semesterIdToSemYear } from '$lib/search/firebase';
+import { error, json } from '@sveltejs/kit';
 
 export async function POST({ request }) {
     const { code, semesterId } = await request.json();
+    const { year, semesterNum } = semesterIdToSemYear(semesterId);
+    const mod = await getModuleDoc(code, semesterNum, year);
 
-    const modulesCollection = collection(db, COLL_MODULES);
-    const modulesQuery = query(
-        modulesCollection,
-        where('code', '==', code),
-        where('semester', '==', semesterId)
-    );
-
-    const querySnapshot = await getDocs(modulesQuery);
-    const module = querySnapshot.docs.map((doc) => doc.data())[0];
-    return json(module);
+    if (mod === undefined) return error(400);
+    return json({ mod });
 }
